Editor: Added a fixup redirectors option that *doesn't* delete the redirectors afterwards, which can be useful for cleaning up a project while others are still actively working and might have created new references to the renamed object #jira UE-143408 #rb lauren.barnes #lockdown julien.marchand #preflight 62215f872c722ee170e2ac96
#ROBOMERGE-AUTHOR: michael.noland
#ROBOMERGE-SOURCE: CL 19258824 in //UE5/Release-5.0/... via CL 19259128
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v924-19243027)

[CL 19263827 by michael noland in ue5-main branch]

+++ Engine/Source/Developer/AssetTools/Private/AssetFixUpRedirectors.cpp
@@ -44,7 +44,7 @@ struct FRedirectorRefs
 };
 
 
-void FAssetFixUpRedirectors::FixupReferencers(const TArray<UObjectRedirector*>& Objects, const bool bCheckoutDialogPrompt) const
+void FAssetFixUpRedirectors::FixupReferencers(const TArray<UObjectRedirector*>& Objects, const bool bCheckoutDialogPrompt, ERedirectFixupMode FixupMode) const
 {
 	// Transform array into TWeakObjectPtr array
 	TArray<TWeakObjectPtr<UObjectRedirector>> ObjectWeakPtrs;
@@ -61,18 +61,18 @@ void FAssetFixUpRedirectors::FixupReferencers(const TArray<UObjectRedirector*>& 
 		{
 			// Open a dialog asking the user to wait while assets are being discovered
 			SDiscoveringAssetsDialog::OpenDiscoveringAssetsDialog(
-				SDiscoveringAssetsDialog::FOnAssetsDiscovered::CreateSP(this, &FAssetFixUpRedirectors::ExecuteFixUp, ObjectWeakPtrs, bCheckoutDialogPrompt)
+				SDiscoveringAssetsDialog::FOnAssetsDiscovered::CreateSP(this, &FAssetFixUpRedirectors::ExecuteFixUp, ObjectWeakPtrs, bCheckoutDialogPrompt, FixupMode)
 				);
 		}
 		else
 		{
 			// No need to wait, attempt to fix references now.
-			ExecuteFixUp(ObjectWeakPtrs, bCheckoutDialogPrompt);
+			ExecuteFixUp(ObjectWeakPtrs, bCheckoutDialogPrompt, FixupMode);
 		}
 	}
 }
 
-void FAssetFixUpRedirectors::ExecuteFixUp(TArray<TWeakObjectPtr<UObjectRedirector>> Objects, const bool bCheckoutDialogPrompt) const
+void FAssetFixUpRedirectors::ExecuteFixUp(TArray<TWeakObjectPtr<UObjectRedirector>> Objects, const bool bCheckoutDialogPrompt, ERedirectFixupMode FixupMode) const
 {
 	TGuardValue<bool> Guard(bIsFixupReferencersInProgress, true);
 
@@ -133,8 +133,11 @@ void FAssetFixUpRedirectors::ExecuteFixUp(TArray<TWeakObjectPtr<UObjectRedirecto
 				// Wait for package referencers to be updated
 				UpdateAssetReferencers(RedirectorRefsList);
 
-				// Delete any redirectors that are no longer referenced
-				DeleteRedirectors(RedirectorRefsList, FailedToSave);
+				if (FixupMode == ERedirectFixupMode::DeleteFixedUpRedirectors)
+				{
+					// Delete any redirectors that are no longer referenced
+					DeleteRedirectors(RedirectorRefsList, FailedToSave);
+				}
 
 				// Finally, report any failures that happened during the rename
 				ReportFailures(RedirectorRefsList);
